Summary: Fixed PAR DMID issues; finish outputCandidate; can now assemble
